Changed Marketplace to menu
renamed the marketplace class to menu because it just contains the menu. I created a separate marketplace class which will have methods for dealing with items/listing that the menu can interact with. I just did this to make it easier to split up work, I can work on the new marketplace class.